design2code-api/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
google.generativeai